+++ q_compress/src/bit_writer.rs
@@ -1,15 +1,17 @@
+use crate::bits;
+use crate::bits::BASE_BIT_MASK;
 use crate::errors::{QCompressError, QCompressResult};
 use crate::data_types::UnsignedLike;
 use crate::constants::{BITS_TO_ENCODE_N_ENTRIES, BYTES_PER_WORD, MAX_ENTRIES, WORD_SIZE};
 
-/// `BitWriter` builds a `Vec<usize>`, enabling a compressor to write bit-level
-/// information and ultimately output a `Vec<u8>`.
+/// Builds compressed data, enabling a [`Compressor`][crate::Compressor] to
+/// write bit-level information and ultimately output a `Vec<u8>`.
 ///
-/// It does this by maintaining a bit index from 0-usize::BITS within its most
-/// recent byte.
+/// It does this by maintaining a bit index from 0 to `usize::BITS` within its
+/// most recent `usize`.
 ///
-/// The reader is consider is considered "aligned" if the current bit index
-/// is 0 or usize::BITS (i.e. at the start or end of the current byte).
+/// The writer is consider is considered "aligned" if the current bit index
+/// is byte-aligned; e.g. `bit_idx % 8 == 0`.
 #[derive(Clone)]
 pub struct BitWriter {
   words: Vec<usize>,
@@ -75,7 +77,7 @@ impl BitWriter {
     self.refresh_if_needed();
 
     if b {
-      *self.last_mut() |= 1_usize << (WORD_SIZE - 1 - self.j);
+      *self.last_mut() |= BASE_BIT_MASK >> self.j;
     }
 
     self.j += 1;
@@ -143,7 +145,7 @@ impl BitWriter {
   }
 
   pub(crate) fn finish_byte(&mut self) {
-    self.j = ((self.j + 7) / 8) * 8;
+    self.j = bits::ceil_div(self.j, 8) * 8;
   }
 
   pub(crate) fn overwrite_usize(&mut self, bit_idx: usize, x: usize, n: usize) {
@@ -170,14 +172,8 @@ impl BitWriter {
   /// Returns the bytes produced by the writer.
   pub fn bytes(&self) -> Vec<u8> {
     let byte_size = self.byte_size();
-    // We can't just transmute because many machines are little-endian.
-    let mut res = self.words.iter()
-      .map(|w| w.to_be_bytes())
-      .flatten()
-      .collect::<Vec<_>>();
-    unsafe {
-      res.set_len(byte_size);
-    }
+    let mut res = bits::words_to_bytes(&self.words);
+    res.truncate(byte_size);
     res
   }
 }