+++ q_compress/src/errors.rs
@@ -45,6 +45,21 @@ impl QCompressError {
   pub(crate) fn insufficient_data<S: AsRef<str>>(message: S) -> Self {
     Self::new(ErrorKind::InsufficientData, message)
   }
+  
+  pub(crate) fn insufficient_data_recipe(
+    name: &str,
+    bits_to_read: usize,
+    bit_idx: usize,
+    total_bits: usize,
+  ) -> Self {
+    Self::insufficient_data(format!(
+        "{}: cannot read {} bits at bit idx {} out of {}",
+        name,
+        bits_to_read,
+        bit_idx,
+        total_bits,
+    ))
+  }
 
   pub(crate) fn invalid_argument<S: AsRef<str>>(message: S) -> Self {
     Self::new(ErrorKind::InvalidArgument, message)